Reorganized file structure and added navbar styles
