+++ src/styles/header.css
@@ -1,0 +1,3 @@
+.header {
+  background-color: rgb(217, 209, 224);
+}
